Moved ButtonDetector to utils folder
